+++ src/core/embed/hexText.ts
